+++ platform/lib/patchbay_web/endpoint.ex
@@ -60,5 +60,6 @@ defmodule PatchbayWeb.Endpoint do
   plug Plug.MethodOverride
   plug Plug.Head
   plug Plug.Session, @session_options
+  plug PatchbayWeb.Plugs.ReadBudget
   plug PatchbayWeb.Router
 end